PySide6: Fix build when defining QT_NO_CAST_FROM_ASCII for the modules
This unearthes a few conversion errors.

Pick-to: 6.6

+++ sources/pyside6/PySide6/glue/qtcore.cpp
@@ -1440,7 +1440,7 @@ auto res = (*%CPPSELF) + cppArg0;
 
 // @snippet conversion-pystring-char
 char c = %CONVERTTOCPP[char](%in);
-%out = %OUTTYPE(c);
+%out = %OUTTYPE(static_cast<unsigned short>(c));
 // @snippet conversion-pystring-char
 
 // @snippet conversion-pyint
